# docs/openapi.yaml
openapi: 3.0.3
info:
  title: flowctl HTTP API
  version: "1"
  description: |
    HTTP API for [flowctl](https://github.com/cvhariharan/flowctl), a single-binary
    self-service workflow platform.

    ## Authentication

    Two mechanisms are supported. **Browser clients** authenticate via session
    cookies (`POST /login` or the OIDC flow) and must include the `X-CSRF-Token`
    header on every state-changing request. **Programmatic clients** authenticate
    via user-scoped API tokens minted in the web UI and sent in the
    `Authorization: Bearer <token>` header (token prefix `fctl_pat_`). Bearer
    requests are exempt from CSRF.

    API-token requests **cannot** create or revoke API tokens themselves, those
    endpoints require a session.

    ## Conventions

    - All resource IDs are UUIDs.
    - Timestamps are RFC 3339 (`2025-06-04T10:00:00Z`).
    - Errors use a uniform JSON envelope (see `Error` schema).
    - Pagination endpoints return `page_count` and `total_count` next to the page slice.
    - Request bodies must be `application/json`, except for flow triggers and flow import
      which take `multipart/form-data` or `application/x-www-form-urlencoded`. Any other
      content type is rejected with `415`, on every endpoint and every method.
servers:
  - url: https://flowctl.example.com
    description: Production server (replace with your deployment URL)
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/v1/{namespace}/flows/{flowID}/schedules:
    parameters:
      - $ref: "#/components/parameters/NamespacePath"
      - { name: flowID, in: path, required: true, schema: { type: string } }
    get:
      operationId: ListSchedules
      tags: [Schedules]
      summary: List user-created schedules for a flow
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/CountPerPage"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SchedulesPaginateResponse" }
    post:
      operationId: CreateSchedule
      tags: [Schedules]
      summary: Create a schedule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduleCreateReq" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduleResp" }
  /api/v1/{namespace}/flows/{flowID}/schedules/{schedule_id}:
    parameters:
      - $ref: "#/components/parameters/NamespacePath"
      - { name: flowID, in: path, required: true, schema: { type: string } }
      - { name: schedule_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      operationId: GetSchedule
      tags: [Schedules]
      summary: Get a schedule
      description: >
        The schedule must belong to the flow named in the path; a schedule_id belonging to a
        different flow is reported as not found.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduleResp" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: UpdateSchedule
      tags: [Schedules]
      summary: Update a schedule
      description: >
        The schedule must belong to the flow named in the path. Only the schedule's creator, a
        namespace admin, or a superuser may update it; other callers get 404.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduleUpdateReq" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduleResp" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: DeleteSchedule
      tags: [Schedules]
      summary: Delete a schedule
      description: >
        The schedule must belong to the flow named in the path. Only the schedule's creator, a
        namespace admin, or a superuser may delete it; other callers get 404.
      responses:
        "200": { description: Deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------- Nodes ----------------
...
  /api/v1/permissions/check:
    post:
      operationId: CheckPermission
      tags: [Permissions]
      summary: Check a specific permission for the calling user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: |
                Implementation accepts a free-form object with the resource and
                action to check. Refer to `internal/handlers/auth.go` for the
                exact shape — typically `{ "resource": "...", "action": "...", "domain": "..." }`.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "fctl_pat_<random>"
      description: |
        User-scoped API token. Mint via `POST /api/v1/users/me/api-tokens` while
        logged in via the browser. Send as `Authorization: Bearer fctl_pat_...`.
    executorBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "fctl_<executor_name>.<signature>"
      description: |
        Executor token generated by the server for a running executor. Send as
        `Authorization: Bearer fctl_<executor_name>.<signature>`. Executors may
        also send `X-User-UUID` to associate requests with the triggering user.
    cookieAuth:
      type: apiKey
      in: cookie
      name: flowctl_session
      description: |
        Session cookie set by `POST /login` or the OIDC flow. Browser clients
        must also send `X-CSRF-Token` on state-changing requests.
  parameters:
    NamespacePath:
      name: namespace
      in: path
      required: true
      schema: { type: string }
      description: Namespace name (slug, not UUID).
...
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 1 }
      description: 1-based page number; `0` means the first page.
    CountPerPage:
      name: count_per_page
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 10 }
  responses:
...
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable message.
        code:
          type: string
          description: Stable machine code.
          enum:
            - VALIDATION_FAILED
            - REQUIRED_FIELD_MISSING
            - INVALID_INPUT
            - INVALID_PAGINATION
            - AUTHENTICATION_FAILED
            - INVALID_CREDENTIALS
            - UNAUTHORIZED
            - FORBIDDEN
            - RESOURCE_NOT_FOUND
            - OPERATION_FAILED
            - INTERNAL_ERROR
        details:
          description: Optional, code-specific payload.

    # Auth
...
    ScheduleCreateReq:
      type: object
      required: [cron, timezone, inputs]
      properties:
        name: { type: string, maxLength: 150 }
        cron: { type: string, description: Standard cron expression. }
        timezone: { type: string, description: IANA timezone name. }
        inputs:
          type: object
          additionalProperties: true
    ScheduleUpdateReq:
      type: object
      required: [cron, timezone, inputs]
      properties:
        name:
          type: string
          maxLength: 150
          description: Omit to keep the current name; send an empty string to clear it.
        cron: { type: string }
        timezone: { type: string }
        inputs:
          type: object
          additionalProperties: true
        is_active: { type: boolean }
    ScheduleResp:
      type: object
      properties:
        uuid: { type: string, format: uuid }
        flow_slug: { type: string }
        flow_name: { type: string }
        name: { type: string, maxLength: 150 }
        cron: { type: string }
        timezone: { type: string }
        inputs:
          type: object
          additionalProperties: true
        created_by: { type: string }
        is_active: { type: boolean }
        is_user_created: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SchedulesPaginateResponse:
      type: object
      properties:
        schedules:
          type: array
          items: { $ref: "#/components/schemas/ScheduleResp" }
        page_count: { type: integer, format: int64 }
        total_count: { type: integer, format: int64 }

    # Nodes
...